+++ libs/schemas/src/events/index.ts
@@ -44,9 +44,6 @@ export enum EventNames {
 
   // Referrals
   SignUpFlowCompleted = 'SignUpFlowCompleted',
-
-  //UnSubscribed email notifications
-  UnSubscribeEmail = 'UnSubscribeEmail',
 }
 
 export type EventPairs =
@@ -157,8 +154,4 @@ export type EventPairs =
   | {
       event_name: EventNames.SignUpFlowCompleted;
       event_payload: z.infer<typeof events.SignUpFlowCompleted>;
-    }
-  | {
-      event_name: EventNames.UnSubscribeEmail;
-      event_payload: z.infer<typeof events.SubscriptionPreferencesUpdated>;
     };
